gyp: run gyp-mac-tool with Python from find-python
Ensures the generated Makefile executes gyp-mac-tool with a Python
that's validated by node-gyp's lib/find-python.js...
unless the Python at that path no longer exists, in which case
fall back to whatever Python is indicated by the shebang line.

Before this, the first `python`/`python3` on the PATH was always used
(based on the shebang line `#!/usr/bin/env python[3]` in gyp-mac-tool)
which might or might not have been a new enough version of Python.

(Pythons validated by lib/find-python.js are always new enough,
as they have passed a semver check tailored to the language features
we use, as well as tailored to which versions of Python
are still officially supported.)