ci: remove incorrect quoting
Also: make sure this is run in bash. The particulars of parameter
expansion and quoting are important in this case. (We don't want
"npx -q some_module" to be read as one command, with spaces in the name
and no arguments. That would result in a "command not found" error.)