feat(lib): try py launcher before brute checking paths in find-python
This is for Windows only.

in lib/find-python.js, when env vars NODE_GYP_FORCE_PYTHON and PYTHON,
CLI flag --python, and `python3` and `python` on the PATH don't give
a usable Python install, check the py.exe launcher
*before* exhaustively checking all the default install paths
that could be used by the python.org Python installer for Windows.

This commit does not *add* or *modify* any checks per se,
it only *reorders* the checks so that py.exe is checked before
the exhaustive path checks happen.

Again, this only applies to Windows. Neither of the checks this commit
reorders are done on OSes other than Windows.